LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/prefetch-uses-cache.sub.https_same-site-expected.txt


FAIL Test that prefetches use/store responses to/from the HTTP cache. assert_equals: Navigation should have retrieved the response from the HTTP Cache. expected "cache" but got "navigational-prefetch"
